order test

+++ src/test/java/com/tsyrulik/dmitry/model/dao/impl/ClientDAOImplTest.java
@@ -36,18 +36,18 @@ public class ClientDAOImplTest {
         ConnectionPool.getInstance(URL, USERNAME, PASSWORD, poolSize);
 
         ConnectionPool.getInstance(1);
-        examplefood = new Food((long) 7, "Шоколад 100 гр", LocalDate.of(2017, 9, 20),
+        examplefood = new Food(7, "Шоколад 100 гр", LocalDate.of(2017, 9, 20),
                 LocalTime.of(15, 0, 0));
-        exampleExercises = new Exercises((long) 6, "широчайшие мышцы спины",
+        exampleExercises = new Exercises(6, "широчайшие мышцы спины",
                 "тяга верхнего блока перед собой; тяга гантели одной рукой; тяга гантели к поясу; отжимания стоя на руках; подтягивания широким хватом",
                 "гантели, турник, тренажер");
 
         allClients.add(new Client(2, "Pety", "Saplov", 23, "M",
-                "goodmail@gmail.com", "58bad6b697dff48f4927941962f23e90", "client", (long) 1,
-                (double) 15, (long) 2));
+                "goodmail@gmail.com", "58bad6b697dff48f4927941962f23e90", "client",  1,
+                (double) 15,  2));
         allClients.add(new Client(4, "Danila", "Letov", 21, "M",
-                "letov@gmail.com", "6982e45352af5526754d83df2d1635", "client", (long) 2,
-                (double) 0.65, (long) 4));
+                "letov@gmail.com", "6982e45352af5526754d83df2d1635", "client", 2,
+                (double) 0.65,  4));
     }
 
 //    @Test
@@ -73,7 +73,7 @@ public class ClientDAOImplTest {
     @Test
     public void CreateAndDeleteClientTest() throws Exception {
         Client client = new Client(4, "Kirill", "Pavlov", 23, "M",
-                "vavl@gmail.com", "58bad6b697dff48f4927941962f23e90", "3", (long) 2, (double) 20, (long) 4);
+                "vavl@gmail.com", "58bad6b697dff48f4927941962f23e90", "3",  2, (double) 20,  4);
         allClients.add(client);
         Client expectedClient = (Client) allClients.get(2);
         clientDAO.createClient(client);
@@ -85,11 +85,11 @@ public class ClientDAOImplTest {
     @Test
     public void updateClientTest() throws DAOFitnessException {
         Client expectedClient = new Client(2, "PetyPety", "SaplovSaplov", 32, "F",
-                "goodmail@gmail.com", "58bad6b697dff48f4927941962f23e90", "3", (long) 1, (double) 15, (long) 2);
+                "goodmail@gmail.com", "58bad6b697dff48f4927941962f23e90", "3", 1, (double) 15,  2);
         Client actualClient = clientDAO.updateClient(expectedClient);
         Assert.assertEquals(actualClient, expectedClient);
         Client expectedClientTwo = new Client(2, "Pety", "Saplov", 23, "M",
-                "goodmail@gmail.com", "58bad6b697dff48f4927941962f23e90", "3", (long) 1, (double) 15, (long) 2);
+                "goodmail@gmail.com", "58bad6b697dff48f4927941962f23e90", "3",  1, (double) 15, 2);
         Client actualClientTwo = clientDAO.updateClient(expectedClientTwo);
         Assert.assertEquals(actualClientTwo, expectedClientTwo);
     }
